Change the structure of the application
Code of components of the blog, the user, the auth has been moved to the core folder.
All controllers, console commands are collected in the presentation folder.
The vendor classes required for controllers and console commands have been moved
to an infrastructure subfolder in the presentation folder.
In the infrastructure root folder will store the vendor classes for
the entire application.